requirements.txt
fastapi==0.68.0
fastapi-pagination==0.9.3
uvicorn==0.17.6
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-multipart==0.0.5
lxml==4.8.0
PyYAML==6.0
pydantic==1.9.0
starlette==0.14.2
